reorganize scripts
Organize scripts into a directory and use startup.m to add the scripts
to MATLAB path. Thus, one can run the experiments directly after opening
MATLAB in the repository root.

Updated versions of the experiments are added to better organize the
plot generation, and to accomodate the new "optimal" choice of base `a`.